fix(agents): preserve automatic model fallback semantics
